add function tovalidate image and content

+++ src/Controllers/ItemsController.php
@@ -29,26 +29,21 @@ Class ItemsController{
         if($csrf_token==$_SESSION['csrf_token']){
             $userID=$_SESSION['id_user'];
             if($_FILES["photo"]['name']!=''){
-                $fileType = $_FILES['photo']['type'];
-                $fileSize = $_FILES['photo']['size'];
-                if(in_array($fileType,$typeFileTable)){
-                    if($fileSize<=10000){
-                        $targetDir = "assets/uploads/"; 
-                        $imageName=date("Y_m_d_H_i_s"). basename($_FILES["photo"]["name"]);
-                        $targetFile = $targetDir.$imageName;
-                        move_uploaded_file($_FILES["photo"]["tmp_name"], $targetFile);
-                    } else {
-                        $_SESSION["errorMessage"]="The uploaded image size is too large. Please upload an image with a smaller size.";
-                        $error=false;
-                    }
+                $resultat=$this->validateImage($_FILES['photo'],10000,$typeFileTable);
+                if($resultat){
+                    $targetDir = "assets/uploads/"; 
+                    $imageName=date("Y_m_d_H_i_s"). basename($_FILES["photo"]["name"]);
+                    $targetFile = $targetDir.$imageName;
+                    move_uploaded_file($_FILES["photo"]["tmp_name"], $targetFile);
                 }else {
-                    $_SESSION["errorMessage"]="Invalid file type. Please upload an image with the format JPG, PNG, or JPEG.";
-                    $error=false;
+                    $views->userItemsAdmin();
+                    die();
                 }
             }else{
                 $imageName='deafult.jpeg';
             }
-            if($title !="" && $content!="" && $category !='null' && $error==true){
+            $result=$this->validateInput($title,$content,$category);
+            if($result && $error==true){
                 $title = htmlspecialchars($title, ENT_QUOTES,'UTF-8');
                 $content = htmlspecialchars($content, ENT_QUOTES,'UTF-8');
                 $item = new Item($title,$content,$userID,$category,$imageName);
@@ -60,13 +55,8 @@ Class ItemsController{
                     }
                 }
                 $_SESSION['succesAdd']="ok";
-                $views->userItemsAdmin();
-            }else{
-                if($title =="") $_SESSION["errorMessage"]="Title is required. Please provide a title for the Item.";
-                if($content =="") $_SESSION["errorMessage"]="Content is required. Please provide a content for the Item.";
-                if($category =="null") $_SESSION["errorMessage"]="Category is required. Please select a category for the Item.";
-                $views->userItemsAdmin();
             }
+            $views->userItemsAdmin();
         }
     }
     public function deletItemUser(){
@@ -91,31 +81,47 @@ Class ItemsController{
     public function updateItem(){
         AuthMiddlewareController::handle();
         extract($_POST);
-        $userID=$_SESSION['id_user'];
-        if($_FILES["photo"]['name']!=''){
-            $imagePath = 'assets/uploads/'.$urlimage;
-            if (file_exists($imagePath) && $urlimage != "deafult.jpeg") {
-                unlink($imagePath);
-            } 
-
-            $targetDir = "assets/uploads/"; 
-            $imageName=date("Y_m_d_H_i_s"). basename($_FILES["photo"]["name"]);
-            $targetFile = $targetDir.$imageName;
-            move_uploaded_file($_FILES["photo"]["tmp_name"], $targetFile);
-        }else{
-            $imageName=$urlimage;
-        }
-        $item = new Item($title,$content,$userID,$category,$imageName);
-        $item->updateItem($idWiki);
-        if(!empty($Tags) && count($Tags)>0){
-            WikiTags::deletWikTags($idWiki);
-            for($i= 0;$i<count($Tags);$i++){   
-                $tagWiki = new WikiTags($idWiki,$Tags[$i]);
-                $tagWiki->addWikiTags();
-            }
-        }
+        $error=true;
         $views = new HomeController();
-        $views->userItemsAdmin();
+        unset($_SESSION["errorMessage"]);
+        $typeFileTable=['image/jpeg','image/png','image/jpg'];
+        if($csrf_token==$_SESSION['csrf_token']){
+            $userID=$_SESSION['id_user'];
+            $resultat=$this->validateImage($_FILES['photo'],10000,$typeFileTable);
+            if($_FILES["photo"]['name']!=''){
+                if($resultat){
+                    $imagePath = 'assets/uploads/'.$urlimage;
+                    if (file_exists($imagePath) && $urlimage != "deafult.jpeg") {
+                        unlink($imagePath);
+                    } 
+                    $targetDir = "assets/uploads/"; 
+                    $imageName=date("Y_m_d_H_i_s"). basename($_FILES["photo"]["name"]);
+                    $targetFile = $targetDir.$imageName;
+                    move_uploaded_file($_FILES["photo"]["tmp_name"], $targetFile);
+                }else {
+                    $views->userItemsAdmin();
+                    die();
+                }
+            }else{
+                $imageName="deafult.jpeg";
+            }
+            $result=$this->validateInput($title,$content,$category);
+            if($result && $error==true){
+                $title = htmlspecialchars($title, ENT_QUOTES,'UTF-8');
+                $content = htmlspecialchars($content, ENT_QUOTES,'UTF-8');
+                $item = new Item($title,$content,$userID,$category,$imageName);
+                $item->updateItem($idWiki);
+                if(!empty($Tags) && count($Tags)>0){
+                    WikiTags::deletWikTags($idWiki);
+                    for($i= 0;$i<count($Tags);$i++){   
+                        $tagWiki = new WikiTags($idWiki,$Tags[$i]);
+                        $tagWiki->addWikiTags();
+                    }
+                }
+                $_SESSION['succesAdd']="ok";
+            }
+            $views->userItemsAdmin();
+        }
     }
     public function archiveItem(){
         AuthMiddlewareController::handle();
@@ -127,5 +133,35 @@ Class ItemsController{
         $data['oneItem']=Item::searchItemById($id);
         $data['wikis']=WikiTags::getWikisTags($data['oneItem']['wikiID']);
         Controller::render('user/details', $data);
+    }
+    public static function validateImage($file, $maxSize, $allowedTypes) {
+        $test=true;
+        $fileType = $file['type'];
+        $fileSize = $file['size'];
+        if (!in_array($fileType, $allowedTypes)) {
+            $_SESSION["errorMessage"] = 'Invalid file type. Please upload an image with the format JPG, PNG, or JPEG.';
+            $test = false;
+        } 
+        if ($fileSize >= $maxSize) {
+            $_SESSION["errorMessage"] = 'The uploaded image size is too large. Please upload an image with a smaller size.';
+            $test = false;
+        }
+        return $test;
+    }
+    public static function validateInput($title, $content, $category) {
+        $test=true;
+        if($title ==""){ 
+            $_SESSION["errorMessage"]="Title is required. Please provide a title for the Item.";
+            $test = false;
+        }
+        if($content ==""){
+            $_SESSION["errorMessage"]="Content is required. Please provide a content for the Item.";
+            $test = false;
+        } 
+        if($category =="null") {
+            $_SESSION["errorMessage"]="Category is required. Please select a category for the Item.";
+            $test = false;
+        }
+        return $test;
     }
 }
